build: clear more assorted type, comment, and literal errors
- CDataScanner: reword the block-comment describing a C comment (no nested slash-
  star), and drop __autoreleasing from the const void ** out-pointer.
- ChallengeModeRootView: type packed NSMutableData * for decipher:; cast the tag to
  int for setDetailInfo:; type tuneInfo TuneInfo *; (void) the parallel-but-unused
  smallFont.
- EditSequence: annotate the intentional switch fall-through.
- EditNoteRendererPad: cast endPosition unsigned for the rightDot comparison.
- EditViewController: apply the kPasteVisibleMargin and kPhoneButtonSizeBase
  constants at their intended sites (both 80/100, previously literals/aliases).
- ChallengeRankingListView: drop the dead kConfirmButtonIndex (the binary reads the
  alert button message and discards it, branching only on the tag).
- EditDataManager: the default chart/editor names are the Japanese UTF-16 strings
  "新しい譜面"/"作成者", not the mis-transcribed raw-byte escapes.

+++ Project/EditNoteRendererPad.m
@@ -788,7 +788,7 @@ static const float kGridZoomMax = 1.0f;
                     break;
                 }
             }
-            if (ev->endPosition > rightDot) {
+            if ((unsigned int)ev->endPosition > rightDot) {
                 break;
             }
             ++ev;
